Revamped Spawning System and Integrated Custom Inspector for Beatmap Obj
Game is able to spawn strings of beats that simulate a Slider in Osu. Added a Custom Inspector that only requires the Folder Name of the Beatmap and Audio file to be manually added. Custom Inspector has a Button to initialise all the maps in the folder and sort them via difficulty.